+++ front-end/playground/src/pages/profile/Profile.css
@@ -1,12 +1,33 @@
-.profile-title {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 16px;
+.profile-page-container {
+    flex-grow: 1;
     display: flex;
-    flex-direction: row;
-    justify-content: space-between;
+    flex-wrap: wrap;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: flex-start;
 }
 
-.lowerInput {
-    margin-top: 1rem;
+.profile-page-nav {
+    width: 100%;
+    justify-content: center;
+    
+}
+.profile-page-nav a{
+    color: black !important;
+    cursor: pointer !important;
+}
+.profile-page-nav a:hover{
+    color: #ed1b76 !important;
+}
+.profile-page-nav .active{
+    color: #ed1b76 !important;
+    border-bottom: 2px solid
+}
+
+.custom-tab-content-pp-general {
+    width: 100%;
+    flex-grow: 1;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
 }
