+++ front-end/playground/src/pages/profile/Tabs/PersonalInfo/PersonalInfo.css
@@ -1,0 +1,12 @@
+.profile-title {
+    font-size: 24px;
+    font-weight: bold;
+    margin-bottom: 16px;
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
+
+.lowerInput {
+    margin-top: 1rem;
+}
